Fix ordering of attached events
Attached events should resolve before the post-handler functions
associated with an event. Thus, the interrupt window needs to open
immediately after the handler for the event.
