+++ server/game/gamesteps/BaseAbilityWindow.js
@@ -1,8 +1,6 @@
 const uuid = require('uuid');
 
-const BaseStep = require('./basestep.js');
-const SimultaneousEvents = require('../SimultaneousEvents');
-const InterruptWindow = require('./InterruptWindow');
+const BaseStep = require('./basestep');
 
 class BaseAbilityWindow extends BaseStep {
     constructor(game, properties) {
@@ -72,22 +70,7 @@ class BaseAbilityWindow extends BaseStep {
             return;
         }
 
-        let attachedEvents = [];
-        for(let event of this.event.getConcurrentEvents()) {
-            attachedEvents = attachedEvents.concat(event.attachedEvents);
-            event.clearAttachedEvents();
-        }
-
-        if(attachedEvents.length === 0) {
-            return;
-        }
-
-        let groupedEvent = new SimultaneousEvents();
-        for(let attachedEvent of attachedEvents) {
-            groupedEvent.addChildEvent(attachedEvent);
-        }
-
-        this.game.queueStep(new InterruptWindow(this.game, groupedEvent, this.postHandlerFunc));
+        this.game.openInterruptWindowForAttachedEvents(this.event);
     }
 }
