+++ test/server/gamesteps/BaseAbilityWindow.spec.js
@@ -130,46 +130,4 @@ describe('BaseAbilityWindow', function() {
             expect(this.window.hasResolvedAbility(this.abilitySpy, this.eventSpy));
         });
     });
-
-    describe('openWindowForAttachedEvents()', function() {
-        describe('when the event has no attached events', function() {
-            beforeEach(function() {
-                this.eventSpy.attachedEvents = [];
-                this.window.openWindowForAttachedEvents();
-            });
-
-            it('should not open an interrupt window', function() {
-                expect(this.gameSpy.queueStep).not.toHaveBeenCalled();
-            });
-        });
-
-        describe('when the event has attached events', function() {
-            beforeEach(function() {
-                this.attachedEvent1 = { event: 1 };
-                this.attachedEvent2 = { event: 1 };
-                this.eventSpy.attachedEvents = [this.attachedEvent1, this.attachedEvent2];
-                this.window.openWindowForAttachedEvents();
-            });
-
-            it('should clear the attached events from the initial event', function() {
-                expect(this.eventSpy.clearAttachedEvents).toHaveBeenCalled();
-            });
-
-            it('should open an interrupt window', function() {
-                const InterruptWindow = require('../../../server/game/gamesteps/InterruptWindow');
-                expect(this.gameSpy.queueStep).toHaveBeenCalledWith(jasmine.any(InterruptWindow));
-            });
-
-            it('should create an event grouping the attached events', function() {
-                expect(this.gameSpy.queueStep).toHaveBeenCalledWith(jasmine.objectContaining({
-                    event: jasmine.objectContaining({
-                        childEvents: [
-                            this.attachedEvent1,
-                            this.attachedEvent2
-                        ]
-                    })
-                }));
-            });
-        });
-    });
 });
